Refactor: refactor SearchPanel Component
